Backend/app/agents/oder_management_agent/requirements.txt
google-adk>=0.1.0
sqlalchemy>=2.0.0
python-dotenv>=0.19.0
pymysql>=1.1.0 